+++ examples/die_rolls.rs
@@ -14,14 +14,14 @@ fn main() {
     let obs: CategoricalData<u8> = DataOrSuffStat::Data(&rolls);
 
     // Log marginal likelihood of the observed rolls
-    println!("Log P(rolls) = {}", prior.ln_m(&obs));
+    println!("Log P(rolls) = {}", prior.ln_m(obs));
 
     // Posterior predictive probability of the next die roll being 5 given the
     // observed rolls.
-    let pp_5: f64 = prior.pp(&5_u8, &obs);
+    let pp_5: f64 = prior.pp(&5_u8, obs);
     println!("P(y = 5 | rolls) = {pp_5}");
 
     // Draw a sample from the posterior distribution P(θ|x)
-    let weights: Vec<f64> = prior.posterior(&obs).draw(&mut rng);
+    let weights: Vec<f64> = prior.posterior(obs).draw(&mut rng);
     println!("Die weight sample: {weights:?}");
 }